feat: implement scaling for inference deployments (#111)

+++ src/main/java/com/epam/aidial/deployment/manager/dao/entity/deployment/PersistenceScalingStrategy.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 public class PersistenceScalingStrategy {
     @JsonProperty("$type")
     private PersistenceScalingStrategyType type;
-    private double threshold;
+    private int threshold;
 }
